ibdo :: create IBaseRecord
